pyproject.toml
[project]
name = "clever-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jvansant13", email = "jermey.vansant@davinciacademy.org" }
]
requires-python = ">=3.12"
dependencies = [
    "oneroster-api>=0.1.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]

[project.scripts]
clever-sync = "clever_sync:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
